+++ src/main/java/com/assetflow/loan/LoanStatus.java
@@ -1,5 +1,6 @@
 package com.assetflow.loan;
 
+
 public enum LoanStatus {
     RENTED, RETURNED, OVERDUE
 }