+++ src/main/java/com/assetflow/loan/service/LoanService.java
@@ -9,10 +9,15 @@ import com.assetflow.loan.dto.*;
 import com.assetflow.loan.repository.LoanRepository;
 import com.assetflow.member.Member;
 import com.assetflow.member.repository.MemberRepository;
+import com.assetflow.reservation.Reservation;
+import com.assetflow.reservation.ReservationStatus;
+import com.assetflow.reservation.repository.ReservationRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -22,6 +27,7 @@ public class LoanService {
     private final LoanRepository loanRepository;
     private final MemberRepository memberRepository;
     private final AssetItemRepository assetItemRepository;
+    private final ReservationRepository reservationRepository;
 
     @Transactional
     public LoanCreateResponse createLoan(LoanCreateRequest request) {
@@ -30,7 +36,10 @@ public class LoanService {
         AssetItem assetItem = assetItemRepository.findById(request.getAssetItemId())
                 .orElseThrow(() -> new IllegalStateException("존재 하지 않는 자산 품목입니다."));
 
+        // 자산이 대출 가능한 상태인지 검증
         if (assetItem.getAssetItemStatus() == AssetItemStatus.AVAILABLE) {
+            completeReadyReservation(member, assetItem);
+
             assetItem.rentAsset();
 
             Loan loan = new Loan(
@@ -51,16 +60,42 @@ public class LoanService {
         throw new IllegalStateException("대출이 불가 합니다.");
     }
 
+    private void completeReadyReservation(Member member, AssetItem assetItem) {
+        List<Reservation> reservations = getReadyReservationsByAsset(assetItem);
+        if (!reservations.isEmpty()) {
+            Reservation reservation = reservations.get(0);
+            if (!reservation.getMember().getId().equals(member.getId())) {
+                throw new IllegalStateException("해당 요청자는 예약자가 아닙니다.");
+            }
+            reservation.completed();
+        }
+    }
+
+    private List<Reservation> getReadyReservationsByAsset(AssetItem assetItem) {
+        return reservationRepository.findByAssetItemIdAndReservationStatusOrderByReservedAtAsc(
+                assetItem.getId(), ReservationStatus.READY);
+    }
+
     @Transactional
     public LoanReturnResponse returnLoan(Long loanId) {
         Loan loan = loanRepository.findById(loanId)
                 .orElseThrow(() -> new IllegalStateException("해당 대여 기록이 존재하지 않습니다."));
 
         AssetItem assetItem = loan.getAssetItem();
+
         if (loan.getLoanStatus() == LoanStatus.RENTED) {
             loan.returnLoan();
-            assetItem.returnAsset();
 
+            List<Reservation> reservations =
+                    reservationRepository.findByAssetItemIdAndReservationStatusOrderByReservedAtAsc(
+                            assetItem.getId(), ReservationStatus.WAITING);
+
+            if (!reservations.isEmpty()) {
+                Reservation reservation = reservations.get(0);
+                reservation.ready();
+            }
+
+            assetItem.returnAsset();
             return new LoanReturnResponse(
                     loan.getId(),
                     loan.getLoanStatus(),
@@ -68,6 +103,7 @@ public class LoanService {
                     loan.getAssetItem().getId(),
                     loan.getReturnDate()
             );
+
         }
         throw new IllegalStateException("이미 반납된 대여 입니다.");
 
@@ -91,13 +127,23 @@ public class LoanService {
     public List<MyLoanListResponse> findLoansByMember(Long memberId) {
         return loanRepository.findByMemberId(memberId).stream()
                 .map(loan -> new MyLoanListResponse(
-                    loan.getId(),
-                    loan.getLoanStatus(),
-                    loan.getMember().getId(),
-                    loan.getLoanDate(),
-                    loan.getDueDate(),
-                    loan.getReturnDate()
+                        loan.getId(),
+                        loan.getLoanStatus(),
+                        loan.getMember().getId(),
+                        loan.getLoanDate(),
+                        loan.getDueDate(),
+                        loan.getReturnDate()
                 ))
                 .toList();
     }
+
+    @Transactional
+    @Scheduled(cron = "0 0 0 * * *")
+    public void updateOverdueLoan() {
+        List<Loan> overLoans = loanRepository.findByLoanStatusAndDueDateBeforeAndReturnDateIsNull(
+                LoanStatus.RENTED, LocalDate.now()
+        );
+        overLoans.forEach(loan -> loan.markOverdue());
+    }
+
 }